backend_python/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pydantic
python-multipart
Pillow
pytesseract
python-dateutil
regex